add comparison reviewers to validation factories
